Fixed the food being picked up when adjacent to it
Reduced the box colliders for the food and soda to 0.9X and 0.9Y and
made the player box collider's X value 0.8 instead of 0.9. This removed
the bug where food and soda was being picked up upon hitting a tile
adjacent to the food instead of when moving directly onto the food tile.