Added pause screen/function, refactored a tiny bit, and reset the restart prompt to default to Yes
A pause screen now appears over the game when the player presses the
escape key. The game reduces its background music volume and does not
allow any events to occur while the game is paused. I also refactored
the code for the reset prompt upon game over, and I made it default to
"Yes" upon a second or more restarts.