Added a button for pausing on touch screens
The pause button is now available to be pressed to pause the game when
playing on touch screens. At the moment, it only works if I disable the
#if UNITY_EDITOR from the code in GameManager when using the Unity
remote. Need to fix this later. Also, as a result of the current
implementation, the game unpauses when you click on the game screen
while it is paused. This is OK, but probably needs to be disabled when
in the editor or the standalone.